refactor: update import paths to remove file extensions and streamline exports
